Revert "chore: move thread_pool.py to tools"
This reverts commit c94cbebf346ecb2c84755e45ee90f41a436539a5.
